flatten all example app tests
