+++ packages/patrol/example/integration_test/open_app_test.dart
@@ -4,7 +4,7 @@ import 'dart:io' show Platform;
 
 import 'package:flutter/material.dart';
 
-import '../common.dart';
+import 'common.dart';
 
 Future<void> main() async {
   late String mapsId;
