+++ packages/patrol/example/integration_test/open_quick_settings_test.dart
@@ -1,4 +1,4 @@
-import '../common.dart';
+import 'common.dart';
 
 void main() {
   patrol('opens quick settings', ($) async {
