+++ packages/patrol/example/integration_test/service_airplane_mode_test.dart
@@ -1,4 +1,4 @@
-import '../../common.dart';
+import 'common.dart';
 
 void main() {
   patrol('disables and enables airplane mode twice', ($) async {
