+++ packages/patrol/example/integration_test/service_wifi_test.dart
@@ -1,4 +1,4 @@
-import '../../common.dart';
+import 'common.dart';
 
 void main() {
   patrol('disables and enables wifi twice', ($) async {
