+++ packages/patrol/example/integration_test/webview_stackoverflow_test.dart
@@ -1,4 +1,4 @@
-import '../../common.dart';
+import 'common.dart';
 
 Future<void> main() async {
   testWebViewC();
